client_backend/requirements.txt
vocode
python-dotenv
elevenlabs